requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
fpdf2>=2.7.7
pytest>=7.4.0
openai>=1.12.0
anthropic>=0.19.0
pydantic>=2.5.0
